fix(resolver): fix incorrect position info of CannotFindModule Error.
There is an incorrect position of ConnotFindModule Error. When a CannotFindModule error is triggered (e.g., importing a non-existent file), the error message is indicated at x:2(line: x and column: 2), but by convention it should be x:1 or just line x. Now fix it to just output the line number.
```
KCL Error [CannotFindModule]
---> File ./test.k:1
1 |import non_existent_file
```

fix #126