Fix: fix cannot find metadata in konfig base file.  (#500)
Fix: fix cannot find metadata in konfig base file. Update the function `lookup_compile_unit_path. This function takes a file path as input and searches for the nearest "kcl.yaml" file or the nearest "project.yaml" file. If a "kcl.yaml" file is found, it returns the path of the directory containing the file. If a "project.yaml" file is found, it returns the path of the first directory containing a "kcl.yaml" file in that project. If neither of these files are found, it returns an error indicating that the files were not found.`